# Conflicts:

#	JavaFXApplication/src/com/tomcatisbabycat/homepanel/schedule/ScheduleController.java

+++ JavaFXApplication/src/com/tomcatisbabycat/homepanel/Condition/DustController.java
@@ -62,7 +62,7 @@ public class DustController implements Initializable {
 		return time;
 	}
 
-	private int getDust() {
+	private double getDust() {
 		return ss.getDust();
 
 	}
